+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Added :class:`~cryptography.hazmat.primitives.twofactor.hotp.HOTP`.
 * Added :class:`~cryptography.hazmat.primitives.twofactor.totp.TOTP`.
+* Added :class:`~cryptography.hazmat.primitives.ciphers.algorithms.IDEA` support.
 
 0.2.2 - 2014-03-03
 ~~~~~~~~~~~~~~~~~~